requirements.txt
pytest==7.2.1
pytest-xdist==2.1.0
allure-pytest==2.8.22
requests==2.24.0
mysql-connector-python==8.0.22
pytest-playwright==0.3.1
playwright==1.31.1
